chore: rebase onto dev/1.0.0
